Unify logic Math and Ucum logic
- Move UCUM related functions from Math.kt to Ucum.kt
- Math.kt is now accessing UCUM functionalities through Ucum.kt
